problem 9, 11, 12, 13 code refactor

+++ problem12.py
@@ -1,12 +1,10 @@
+import math
+
 if __name__ == '__main__':
-
-    import math
-
-
     def count_factors(num):
         count = 0
         perfect_square = False
-        for divisor in range(1, math.floor(num ** 0.5) + 1):
+        for divisor in range(1, int(math.floor(num ** 0.5)) + 1):
             if num % divisor == 0:
                 count += 1
             if divisor == num ** 0.5:
@@ -16,9 +14,7 @@ if __name__ == '__main__':
             return (count * 2) - 1
         return count * 2
 
-
     tally = 0
-
     answer = 0
     current = 1
     while answer == 0:
